Mostly working, problems with spaces and split
Having issues because splits commands at ' ', so if there is a space cut out, it only adds back in what's before the space. In other things, I have gotten around this by changing spaces to "%20" in the string, so I'll probably do the same with this later. Other parts seem to be working, although I have not tested them extensively yet.